requirements.txt
Django==1.6.3
MySQL-python==1.2.5
South==0.8.3
amqp==1.4.4
anyjson==0.3.3
billiard==3.3.0.16
celery==3.1.9
django-celery==3.1.9
django-extensions==1.2.4
ecdsa==0.10
isodate==0.4.9
jsonschema==2.3.0
kombu==3.0.13
pycrypto==2.6.1
pygeoip==0.2.7
python-dateutil==2.1
python-memcached==1.48
pytz==2013.7
raven==3.5.0
requests==2.0.1
six==1.3.0
wsgiref==0.1.2
PyHamcrest==1.8.0
django-tastypie==0.11.0
mimeparse==0.1.3
paramiko==1.12.0